Record test dependencies with dep
The main twirp library does not have any dependencies, but its tests
and generated code exmaples do.

In order to run tests in a CI environment, we need a manifest listing
our dependencies. This adds that list, but does not commit the vendor
directory.

To avoid confusing the dep tool with false constraints, use overrides.
There's a comment in Gopkg.toml more fully explaining this.