Remove all uses of stdlib log
This changes generated server code to never call the standard
library's log package in case of weird errors. Instead, it calls the
Error hook, except when the error is a failure to write an error
response. In that case, we just silently ignore the write error to
avoid calling the Error hook twice in one request.