Fix bug
Fix bug with go code generation.

The generated api.twirp.go file will not compile if a service contains a method with the same name as the services name. This is because the generated code uses an embedded struct.

This fix uses the embedded structs name when making a call to the method.