When a server panics, flush the error
When a service panics, we write an error into the http.ResponseWriter
and re-raise the panic. If nothing up the http.Handler stack
intercepts that panic, then the write might not get flushed in time,
resulting in a confusing "EOF" error message for the client.

We need to flush the error explicitly. But if only do that, then the
Go HTTP library will switch into chunked transfer encoding mode. If
that happens, the client doesn't know how many bytes to wait for in
the error response, and the panicking server will close the HTTP
response uncleanly, without terminating the chunked transfer response.
As a result, if we just flush, the client _still_ just gets a
confusing "EOF" message, since it's left unsure of whether the chunked
transfer is done.

The right thing is to explicitly set the content length of the
response, which prevents the Go HTTP library from switching to
chunked-transfer, and which informs the client of exactly how many
bytes to expect. Only then can we flush.