Fix code generation for imported message defs from the same go_package
It's possible to import a .proto file which has the same 'option
go_package' value as the target file being generated during a run of
protoc --twirp_out. If the go_package is fully specified in both files
as an import path - for example, as 'option go_package =
"github.com/twitchtv/twirp/example"' then we should take care: we
don't want to import that package in generated code, since doing so
would result in an import loop.